Add UpdateQuery class.
Split code specific to UPDATE query generation into a separate class.
